order cancel

+++ src/main/webapp/WEB-INF/views/cart.jsp
@@ -21,6 +21,15 @@
 	 		    checkboxes[i].checked = source.checked;
 	 		  }
 	 		}
+	 		
+	 		function confirmd() {  
+	 			            var msg = "Decide to delete?";  
+	 			            if (confirm(msg)==true){  
+	 			                return true;  
+	 			            }else{  
+	 			                return false;  
+	 			            }  
+	 			        }  
 		</script>
 		
 		<a href="cart.htm">cart</a> | <a href="order.htm">order</a> | <a href="sales.htm">sales record</a>
@@ -41,7 +50,7 @@
 				<c:out value="quantity: ${prod.quantity}" /> | 
 				<i>$<c:out value="price: ${prod.price}" /></i>
 			<a href="cart/edit/${prod.id }">edit</a>
-			<a href="cart/delete/${prod.id }">delete</a>             
+			<a href="cart/delete/${prod.id }" onclick="return confirmd()">delete</a>             
 			<br>
 			<a href="cart/checkout/${prod.id }">checkout</a>
 			<br>
@@ -49,14 +58,12 @@
 		
 		
 		
-		<input type="checkbox" onClick="selectall(this)" /> Select All<br/>
-		
-		<!-- <label><input name="select all" type="checkbox" value="all"/>select all</label> -->
+	<!--  <input type="checkbox" onClick="selectall(this)" /> Select All<br/>
 		</sf:form>
 		
-				<a href="cart/checkout">checkout</a>
+				<a href="cart/checkout">checkout</a>-->
 		
-		<a href="cart/add/${prod.id }">add</a>
+		<a href="cart/add/${prod.id }">add</a>  
 	
 	</body>
 </html>
